% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_box.R
\name{add_box}
\alias{add_box}
\title{Add a box layer to the chess board}
\usage{
add_box(
  bottom_left,
  top_right,
  size = 1.5,
  linetype = "solid",
  fill = NA,
  col = "red",
  ...
)
}
\arguments{
\item{bottom_left}{square for box bottom left coordinate (eg "e4")}

\item{top_right}{square for box top right coordinate}

\item{size}{Box line size}

\item{linetype}{Box line type}

\item{fill}{Box fill (default = NA (no fill))}

\item{col}{Box line colour (default = "red")}

\item{...}{Further arguments passed to \code{ggplot2::geom_rect()}}
}
\value{
A ggplot2 geom_segment() layer
}
\description{
Add a box layer to the chess board
}
\examples{
plot_fen() + add_box("d4", "e5")
plot_fen() +
  add_box("d4", "e5", fill = "white", alpha=1/2) +
  add_box("g3", "h6", fill="green", alpha=1/2, col="black", linetype="solid")
}
