% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_piece.R
\name{add_piece}
\alias{add_piece}
\title{Add a piece to the current plot}
\usage{
add_piece(piece, square, perspective = "w", piece_scale = 0.85)
}
\arguments{
\item{piece}{piece symbol in FEN notation ("P" = white pawn etc...)}

\item{square}{Square to draw piece on}

\item{perspective}{Board perspective for the added piece (needs to be set to the same as persepctive used with \code{plot_fen()})}

\item{piece_scale}{Piece scaling factor (default = 0.85)}
}
\value{

}
\description{
Add a piece to the current plot
}
\examples{
# Draw empty board and add some pieces
plot_fen("8/8/8/8/8/8/8/8 w KQkq - 0 1") +
  add_piece("q", "c3") +
  add_piece("B", "h8")
}
