% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{Plot an empty chess board}
\usage{
board(perspective = "w", cols = "brown")
}
\arguments{
\item{perspective}{should the board be viewed from the white side or black side. One of "w" or "b" (default = "w")}

\item{cols}{A single character colour theme listed below or a length 2 vector of colours ordered dark, light (default = "brown")
\itemize{
  \item "brown"
  \item "blue"
  \item "blue2"
  \item "green"
  \item "grey"
  \item "pink"
  \item "purple"
  \item "ic"
  }}
}
\value{
A ggplot2 object
}
\description{
Plot an empty chess board
}
