% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colour_lookup}
\alias{colour_lookup}
\title{A list of dark and light square colour options}
\format{
A named list of dark and light square colours

A named list of dark and light square colours
}
\usage{
colour_lookup

colour_lookup
}
\description{
<<<<<<< HEAD
Not intended for use by the user
=======
Not inteded for use by the user

Not inteded for use by the user
>>>>>>> dbbe9d4c4998b90156fc0532bed8bc90d56fb35f
}
\keyword{datasets}
