% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fade-board.R
\name{fade_board}
\alias{fade_board}
\title{Fade the chess board}
\usage{
fade_board(fade = 0.5, fill = "white")
}
\arguments{
\item{fade}{fade ammount between 0 (no fade) and 1 (no board)}

\item{fill}{fade colour (default = "white")}
}
\description{
Fading the board can make annotations stand out more
}
\examples{
plot_fen() +
  fade_board() +
  add_arrow("g1", "f3")
}
