% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-fen.R
\name{fen_to_df}
\alias{fen_to_df}
\title{Convert a FEN string to a dataframe of pieces and positions}
\usage{
fen_to_df(fen)
}
\arguments{
\item{fen}{a character vector FEN of length one}
}
\value{
A dataframe
}
\description{
Convert a FEN string to a dataframe of pieces and positions
}
\examples{
fen_to_df("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1")
}
