% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight-square.R
\name{highlight_squares}
\alias{highlight_squares}
\title{Highlight soecific square of chess board}
\usage{
highlight_squares(
  squares,
  col = "red",
  size = 1.5,
  fill = NA,
  res = 50,
  d = 0.45,
  ...
)
}
\arguments{
\item{col}{colour of highlight circle}

\item{size}{thickness of highlight circle}

\item{fill}{fill colour of highlight circle}

\item{...}{further arguments passed to \code{ggplot2::geom_polygon()}}

\item{square}{square to highlight (example: "e4")}
}
\value{
A ggplot2 geom_polygon() layer
}
\description{
Highlight soecific square of chess board
}
