% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notation_lookup}
\alias{notation_lookup}
\title{A named vector for help when parsing FEN strings}
\format{
A named list of dark and light square colours
}
\usage{
notation_lookup
}
\description{
Not intended for use by the user
}
\keyword{datasets}
