% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-fen.R
\name{parse_fen}
\alias{parse_fen}
\title{Parse a FEN text string}
\usage{
parse_fen(fen)
}
\arguments{
\item{fen}{a character vector FEN of length one}
}
\value{
A named character vector
}
\description{
Parse a FEN text string
}
\examples{
parse_fen("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1")
}
