% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fen.R
\name{plot_fen}
\alias{plot_fen}
\title{Visualise a FEN}
\usage{
plot_fen(
  fen,
  perspective = "w",
  cols = c("#b58863ff", "#f0d9b5ff"),
  piece_scale = 0.85
)
}
\arguments{
\item{fen}{a character vector FEN of length one}

\item{perspective}{view board from white "w" or black "b" perspective (default = "w")}

\item{cols}{Length 2 vector giving colours of dark and light squares (default = c("#b58863ff", "#f0d9b5ff"))}

\item{piece_scale}{scaling factor for piece sizes (default = 0.75)}
}
\value{
A ggplot2 plot object
}
\description{
Visualise a FEN
}
\examples{
plot_fen('Q1b2rk1/p1p2p1p/6Bp/2b5/8/2P5/P1P2PPP/R4RK1 b - - 0 16')
}
