% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pieces.R
\name{plot_pieces}
\alias{plot_pieces}
\title{Plot chess pieces}
\usage{
plot_pieces()
}
\value{
A ggplot2 object
}
\description{
Plot chess pieces
}
\details{
Chess piece SVG design file downloaded from https://commons.wikimedia.org/wiki/File:Chess_Pieces_Sprite.svg

 jurgenwesterhof (adapted from work of Cburnett), CC BY-SA 3.0 <https://creativecommons.org/licenses/by-sa/3.0>, via Wikimedia Commons
}
