% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seven_tag_roster}
\alias{seven_tag_roster}
\title{A tibble containing the PGN seven tag roster}
\format{
A tibble
}
\usage{
seven_tag_roster
}
\description{
Not intended for use by the user
}
\keyword{datasets}
