% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{square_lookup}
\alias{square_lookup}
\title{A named vector for help when plotting chess positions}
\format{
A named vector
}
\usage{
square_lookup
}
\description{
Not intended for use by the user
}
\keyword{datasets}
