% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-arrow.R
\name{add_arrow}
\alias{add_arrow}
\title{Add arrow to chess board visualisation}
\usage{
add_arrow(
  start,
  end,
  arrow_style = grid::arrow(),
  size = 1.5,
  col = "blue",
  ...
)
}
\arguments{
\item{start}{start square for arrow as character (example: "e2")}

\item{end}{end square for arrow as character (example: "e4")}

\item{arrow_style}{an arrow specified by \code{grid::arrow()}}

\item{size}{arrow size (line thickness)}

\item{col}{arrow colour}

\item{...}{further arguments passed to \code{ggplot2::geom_segment()}}
}
\value{
A ggplot2 geom_segment() layer
}
\description{
Add arrow to chess board visualisation
}
