% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{board}
\alias{board}
\title{Plot an empty chess board}
\usage{
board(perspective = "w", cols = c("#b58863ff", "#f0d9b5ff"))
}
\arguments{
\item{perspective}{should the board be viewed from the white side or black side. One of "w" or "b" (default = "w")}

\item{cols}{Length 2 vector giving colours of dark and light squares (default = c("#b58863ff", "#f0d9b5ff"))}
}
\value{
A ggplot2 object
}
\description{
Plot an empty chess board
}
\examples{
board()
}
