% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paths}
\alias{paths}
\title{The chess piece polygon paths}
\format{
A dataframe of chess piece polygon coordinates
}
\usage{
paths
}
\description{
The chess piece polygon paths
}
\keyword{datasets}
