% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paths}
\alias{paths}
\title{The chess piece polygon paths}
\format{
A dataframe of chess piece polygon coordinates
}
\usage{
paths
}
\description{
<<<<<<< HEAD
Not intended for use by the user
=======
Not inteded for use by the user
>>>>>>> dbbe9d4c4998b90156fc0532bed8bc90d56fb35f
}
\keyword{datasets}
