% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgn_to_df.R
\name{pgn_to_df}
\alias{pgn_to_df}
\title{Convert a PGN file to a dataframe}
\usage{
pgn_to_df(pgn, pgn_tags = "seven")
}
\arguments{
\item{pgn}{A raw (un-annotated) PGN text file}

\item{pgn_tags}{PGN tags wanted in the output dataframe. One of
\itemize{
  \item "seven" - The 'Seven tag roster' (default)
  \item "all_present" - ALl tags found in the PGN file
  \item "all" - All known tags (experimental)
  }}
}
\value{

}
\description{
This is experimental. It will not work for annotated PGN files.
http://www.saremba.de/chessgml/standards/pgn/pgn-complete.htm#c8.1.1
}
