% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fen.R
\name{plot_fen}
\alias{plot_fen}
\title{Visualise a FEN}
\usage{
plot_fen(
  fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1",
  perspective = "w",
  palette = "brown",
  piece_scale = 0.85,
  show_coords = TRUE,
  show_fen = FALSE,
  newspaper_spacing = 0.12,
  border_col = NA,
  border_size = 0.5
)
}
\arguments{
\item{fen}{a character vector FEN of length one (defaults to the opening position)}

\item{perspective}{view board from white "w" or black "b" perspective (default = "w")}

\item{palette}{A single character colour theme listed below or a length 2 vector of colours ordered dark, light (default = "brown")
\itemize{
  \item "brown"
  \item "blue"
  \item "blue2"
  \item "green"
  \item "grey"
  \item "pink"
  \item "purple"
  \item "ic"
  \item "news"
  }}

\item{piece_scale}{scaling factor for piece sizes (default = 0.85)}

\item{show_coords}{logical - should the board coordinates be printed? (default = TRUE)}

\item{show_fen}{logical - should the FEN be printed in the plot caption? (default = FALSE)}

\item{newspaper_spacing}{spacing for shading lines if \code{palette = "news"}}

\item{border_col}{Chess board border colour (default = NA (no border))}

\item{border_size}{Chess board border size (default = 0.5)}
}
\value{
A ggplot2 plot object
}
\description{
Visualise a FEN
}
\details{
Chess piece SVG design file downloaded from https://commons.wikimedia.org/wiki/File:Chess_Pieces_Sprite.svg

 jurgenwesterhof (adapted from work of Cburnett), CC BY-SA 3.0 <https://creativecommons.org/licenses/by-sa/3.0>, via Wikimedia Commons
}
\examples{
# Plot the starting position
plot_fen('Q1b2rk1/p1p2p1p/6Bp/2b5/8/2P5/P1P2PPP/R4RK1 b - - 0 16')

# Plot an empty board
plot_fen("8/8/8/8/8/8/8/8 w KQkq - 0 1")
}
